Backend/requirements.txt
python-dotenv==1.0.1
flask==3.0.3
flask-restx==1.3.0
flask_cors==4.0.0
google-generativeai==0.5.2
gunicorn==22.0.0